+++ packages/dataviews/src/field-types/stories/index.story.tsx
@@ -173,8 +173,8 @@ const data: DataType[] = [
 		media: 'https://live.staticflickr.com/7398/9458193857_e1256123e3_z.jpg',
 		mediaWithElements:
 			'https://live.staticflickr.com/7398/9458193857_e1256123e3_z.jpg',
-		array: [ 'item1', 'item2', 'item3' ],
-		arrayWithElements: [ 'item1', 'item2', 'item3' ],
+		array: [ 'United States', 'United Kingdom', 'Spain' ],
+		arrayWithElements: [ 'us', 'uk', 'es' ],
 		notype: 'No type',
 		notypeWithElements: 'No type',
 		priceWithPrefix: '25.99',
@@ -504,9 +504,9 @@ const fields: Field< DataType >[] = [
 		label: 'Array (with elements)',
 		description: 'Help for array with elements.',
 		elements: [
-			{ value: 'item1', label: 'Item 1' },
-			{ value: 'item2', label: 'Item 2' },
-			{ value: 'item3', label: 'Item 3' },
+			{ value: 'us', label: 'United States' },
+			{ value: 'uk', label: 'United Kingdom' },
+			{ value: 'es', label: 'Spain' },
 		],
 	},
 	{