+++ packages/dataviews/CHANGELOG.md
@@ -242,6 +242,7 @@
 - DataViews: Add details form layout validation. [#74996](https://github.com/WordPress/gutenberg/pull/74996)
 - Add new `adaptiveSelect` DataForm control. [#74937](https://github.com/WordPress/gutenberg/pull/74937)
 - DataViews: Consistent rendering of selection checkbox and actions in grid layout. [#75056](https://github.com/WordPress/gutenberg/pull/75056)
+- DataViews: Use labels when rendering a list of options. [#74772](https://github.com/WordPress/gutenberg/pull/74772)
 
 ### Code Quality